+++ changelog.md
@@ -1,8 +1,9 @@
 
 ## TODO
+	MIDI post processing
+		arpeggio
+		after grace
 	implicit notes matching with sheet document
-		match lilyNotation with articulated MIDI (ignored repeats), then unfold repeats on lilyNotation, include D.C., D.S.
-		scale setTempo value in matchWithExactMIDI
 		tremolo type note match
 		setTempo missing issue in Once measure layout
 	staffNotation